pyproject.toml
[project]
name = "expense-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.119.1",
    "fastmcp>=2.12.5",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
